Revise PSO implementation

+++ src/emc/motion/motion.h
@@ -283,13 +283,13 @@ typedef struct emcmot_command_t {
     int ssm_mode;           /* spindle sync motion mode: G33(0) /rigid_tap G33.1(1) */
 
     int pso_enable;	/*set pso_enable: enable->1 disable->0 */
-    int pso_mode;		/*(0) switch pso OFF
-							* (1) switch pso ON
-							* (2) switch pso OFF for pso_ticks(x20ns); then ON
-							* (3) switch pso ON  for pso_ticks(x20ns); then OFF
-							*/
+    int pso_mode;       /*(0) switch pso OFF
+                         * (1) switch pso ON
+                         * (2) switch pso OFF for pso_ticks(x20ns); then ON
+                         * (3) switch pso ON  for pso_ticks(x20ns); then OFF
+                         */
     double pso_pitch;	/*pso_pitch*/
-    double pso_tick;	/*pso_tick*/
+    int pso_tick;	/*pso_tick*/
 
 } emcmot_command_t;
 
@@ -775,7 +775,7 @@ typedef struct emcmot_status_t {
     int pso_req;
     int pso_joint;
     int pso_mode;
-    double pso_tick;
+    int pso_tick;
     EmcPose pso_pos;
     double jpso_pos;		// ack one joint psoition which have velocity
 } emcmot_status_t;
